% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothCV.R
\name{ES.grid}
\alias{ES.grid}
\title{ES.Grid}
\usage{
ES.grid(type, alphrange, betarange, nahead, trainset, testset)
}
\arguments{
\item{type}{Type of moving average. Can be \code{"SES"} or \code{"DES"}}

\item{alphrange}{A vector of parameters for the level component. Can be created using \code{seq}, or by manually specifying a vector.}

\item{betarange}{A vector of parameters for the trend component. Can be created using \code{seq}, or by manually specifying a vector.
Not used if type="SES"}

\item{nahead}{Number of observations to forecast.}

\item{trainset}{A set of univariate time series data. Can be a vector (type double) or a data.table.}

\item{testset}{Same as above.}
}
\value{
A data.table containing all combinations of alpha and beta with their respective MSE and MAPE values.
}
\description{
Grid search for exponential smoothing to find optimal parameters that minimizes errors
with respect to a certain test set.
}
\examples{
ES.Grid(type="SES", alphrange=seq(0.01,1,0.01), betarange=NA, nahead=12, crudenow$Close, crudetest$Close)
ES.Grid(type="DES", alphrange=seq(0.1,1,0.1), betarange=seq(0.1,1,0.1), nahead=12, crudenow$Close, crudetest$Close)
}
