% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothCV.R
\name{esWrapper}
\alias{esWrapper}
\title{esWrapper}
\usage{
esWrapper(trainset, smoothing, nahead = 0)
}
\arguments{
\item{trainset}{A set of univariate time series data. Can be a vector (type double) or a data.table.}

\item{smoothing}{An object of class \code{HoltWinters}}

\item{nahead}{Number of observations to predict.}
}
\value{
A list containing \code{smoothed}
(training set, moving averages, and predicted training values)
and \code{forc} (forecasted values to use against a test set).
}
\description{
Function wrapper for exponential smoothing.
}
\examples{
esWrapper(crudenow$close, HoltWinters(crudenow$close, alpha=0.5, beta=0.5, gamma=F), nahead=10)
}
