% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothCV.R
\name{MA.Grid}
\alias{MA.Grid}
\title{Grid search for single and double moving average}
\usage{
MA.Grid(trainset, testset, start = 2, end, dist = 1, nahead, type)
}
\arguments{
\item{trainset}{A set of univariate time series data. Can be a vector (type double) or a data.table.}

\item{testset}{Same as above.}

\item{start}{Starting value for m (see \code{sma.dt} and \code{dma.dt}).}

\item{end}{Maximum value for m. Must be less than or equal to the length of the training set for SMA,
or less than or equal to the length of the training set divided by two for DMA.}

\item{dist}{Distance between successive values for m.
The vector of m values will be constructed as \code{seq(start,end,dist)}.}

\item{nahead}{Number of observations to forecast.}

\item{type}{Type of moving average. Can be \code{"SMA"} or \code{"DMA"}}
}
\value{
A data.table containing tested values of m, with MSE and MAPE for each value.
}
\description{
Grid search for single and double moving averages to find optimal window value that minimize errors
with respect to a certain test set. Iterates across given range of window values.
}
\examples{
MA.Grid(crudenow$Close, crudetest$Close, start=1, end=30, dist=2, nahead=12, type="SMA")
MA.Grid(crudenow$Close, crudetest$Close, start=1, end=30, dist=2, nahead=12, type="DMA")
}
