% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothCV.R
\name{sma.dt}
\alias{sma.dt}
\title{Function wrapper for single moving averages.}
\usage{
sma.dt(trainset, m, nahead = 0)
}
\arguments{
\item{trainset}{A set of univariate time series data. Can be a vector (type double) or a data.table.}

\item{m}{Number of most recent observations where the moving average will be taken at each point in time.
Should be less than or equal to the length of training set.}

\item{nahead}{Number of observations to predict.}
}
\value{
A list containing \code{smoothed}
(training set, moving averages, and predicted training values)
and \code{forc} (forecasted values to use against a test set).
}
\description{
Function wrapper for single moving averages.
}
\examples{
sma.dt(crudenow$close,5,nahead=10)
}
