% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothCV.R
\name{smooth.acc}
\alias{smooth.acc}
\title{smooth.acc}
\usage{
smooth.acc(smoothresult, againstself = T, testset)
}
\arguments{
\item{smoothresult}{An object from the output of \code{dma.dt}, \code{sma.dt}, or \code{esWrapper}}

\item{againstself}{Whether to test forecast accuracy against the training set or a testing set}

\item{testset}{A set of univariate time series data. Can be a vector (type double) or a data.table.}
}
\value{
A data.table containing MSE and MAPE of the inputted smoothing method.
}
\description{
Accuracy measures for results of certain smoothing methods.
}
\examples{
smooth.acc(sma.dt(crudenow$close,5,nahead=10), againstself=F, crudetest$close)
smooth.acc(es.Wrapper(crudenow$close, HoltWinters(crudenow$close, alpha=0.5, beta=0.5, gamma=F), nahead=10),againstself=T)
}
